requirements.txt
gradio>=4.0.0
openai
anthropic
google-genai
python-dotenv
pdfplumber
pydantic
